pyproject.toml
[project]
name = "knowledge-system"
version = "0.1.0"
description = "personal knowledge management system"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.2",
    # 主要な依存関係をここに記載
    "fastapi>=0.115.12",
    "psycopg2-binary>=2.9.10",
    "python-multipart>=0.0.20",
    "python-slugify>=8.0.4",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.3",
    # 認証・セキュリティ関連
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    "pydantic[email]>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "pytest-xdist>=3.5.0",
    "hypothesis>=6.0.0",
    "mypy>=1.10.0",
    "ruff>=0.4.0",
    "pre-commit>=3.7.0",
    "bandit>=1.7.0",
    "pip-audit>=2.6.0",
]

docs = [
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=2.0.0",
    "sphinx-autodoc-typehints>=2.0.0",
    "myst-parser>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/knowledge_system"]

[tool.ruff]
line-length = 88
indent-width = 4
target-version = "py312"

[tool.ruff.lint]
exclude = [".venv"]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "SIM",    # flake8-simplify
    "RUF",    # Ruff-specific rules
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "PL",     # Pylint
]
ignore = [
    "G004",
    "F821",
    "C401",
    "C408",
    "PERF203",
    "PERF401",
    "PLC2401",  # non-ASCII character in function names (allow Japanese test names)
    "PTH118",   # os.path.join() in sys.path modifications
    "PTH120",   # os.path.dirname() in sys.path modifications
    "PLR2004",  # Magic values in tests
    "RUF002",   # Full-width parentheses in Japanese comments
    "RUF003",   # Full-width parentheses in Japanese comments
    "UP038",    # isinstance with tuple in validation
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["PLR2004", "PLR0913"]  # テストでは magic number とパラメータ数を許可
"backend/app/api/**/*.py" = ["B008", "PLR0913"]  # FastAPI依存関数とAPIパラメータ数を許可
"backend/app/deps.py" = ["B008"]  # FastAPI dependency injection patterns

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
enable_incomplete_feature = ["NewGenericSyntax"]

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false

[tool.bandit]
exclude_dirs = ["tests", ".venv"]
skips = ["B101", "B105", "B106"]  # Skip assert_used test, hardcoded password string, hardcoded password funcarg

[tool.pytest.ini_options]
testpaths = ["tests", "backend/tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--cov-branch",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest-asyncio>=1.0.0",
]
